@@ -17,6 +17,7 @@ import (
 	"bufio"
 	"bytes"
 	"context"
+	"errors"
 	"fmt"
 	"io"
 	"log"
@@ -238,7 +239,7 @@ prefix.name_bucket;constname=constvalue;labelname=val2;le=+Inf 3 1477043
 	got := buf.String()
 
 	if err := checkLinesAreEqual(want, got, useTags); err != nil {
-		t.Fatalf(err.Error())
+		t.Fatal(err.Error())
 	}
 }
 
@@ -290,7 +291,7 @@ prefix.name;constname=constvalue;labelname=val2 1 1477043
 	got := buf.String()
 
 	if err := checkLinesAreEqual(want, got, useTags); err != nil {
-		t.Fatalf(err.Error())
+		t.Fatal(err.Error())
 	}
 }
 
@@ -322,7 +323,7 @@ func checkLinesAreEqual(w, g string, useTags bool) error {
 			log += fmt.Sprintf("want: %v\ngot: %v\n\n", wantSplit, gotSplit)
 
 			if !reflect.DeepEqual(wantSplit, gotSplit) {
-				return fmt.Errorf(log)
+				return errors.New(log)
 			}
 		}
 		return nil
