@@ -263,12 +263,12 @@ func TestSummaryConcurrency(t *testing.T) {
 			ε := objMap[wantQ]
 			gotQ := *m.Summary.Quantile[i].Quantile
 			gotV := *m.Summary.Quantile[i].Value
-			min, max := getBounds(allVars, wantQ, ε)
+			minBound, maxBound := getBounds(allVars, wantQ, ε)
 			if gotQ != wantQ {
 				t.Errorf("got quantile %f, want %f", gotQ, wantQ)
 			}
-			if gotV < min || gotV > max {
-				t.Errorf("got %f for quantile %f, want [%f,%f]", gotV, gotQ, min, max)
+			if gotV < minBound || gotV > maxBound {
+				t.Errorf("got %f for quantile %f, want [%f,%f]", gotV, gotQ, minBound, maxBound)
 			}
 		}
 		return true
@@ -353,12 +353,12 @@ func TestSummaryVecConcurrency(t *testing.T) {
 				ε := objMap[wantQ]
 				gotQ := *m.Summary.Quantile[j].Quantile
 				gotV := *m.Summary.Quantile[j].Value
-				min, max := getBounds(allVars[i], wantQ, ε)
+				minBound, maxBound := getBounds(allVars[i], wantQ, ε)
 				if gotQ != wantQ {
 					t.Errorf("got quantile %f for label %c, want %f", gotQ, 'A'+i, wantQ)
 				}
-				if gotV < min || gotV > max {
-					t.Errorf("got %f for quantile %f for label %c, want [%f,%f]", gotV, gotQ, 'A'+i, min, max)
+				if gotV < minBound || gotV > maxBound {
+					t.Errorf("got %f for quantile %f for label %c, want [%f,%f]", gotV, gotQ, 'A'+i, minBound, maxBound)
 				}
 			}
 		}
@@ -410,20 +410,20 @@ func TestSummaryDecay(t *testing.T) {
 	}
 }
 
-func getBounds(vars []float64, q, ε float64) (min, max float64) {
+func getBounds(vars []float64, q, ε float64) (minBound, maxBound float64) {
 	// TODO(beorn7): This currently tolerates an error of up to 2*ε. The
 	// error must be at most ε, but for some reason, it's sometimes slightly
 	// higher. That's a bug.
 	n := float64(len(vars))
 	lower := int((q - 2*ε) * n)
 	upper := int(math.Ceil((q + 2*ε) * n))
-	min = vars[0]
+	minBound = vars[0]
 	if lower > 1 {
-		min = vars[lower-1]
+		minBound = vars[lower-1]
 	}
-	max = vars[len(vars)-1]
+	maxBound = vars[len(vars)-1]
 	if upper < len(vars) {
-		max = vars[upper-1]
+		maxBound = vars[upper-1]
 	}
 	return
 }
